clean up test files, licenses, proper index
