+++ packages/search/src/users/queries.ts
@@ -1,3 +1,6 @@
+/* Copyright (c) 2021 Environmental Systems Research Institute, Inc.
+ * Apache-2.0 */
+
 import { gql } from "graphql-request";
 
 const userFieldsFragment = gql`
