+++ packages/search/test/users/users.test.ts
@@ -1,3 +1,6 @@
+/* Copyright (c) 2021 Environmental Systems Research Institute, Inc.
+ * Apache-2.0 */
+
 import {
   UserService,
   ISearchUsersFilter,
